+++ samples/FSharp.CosmosDb.Samples/Program.fs
@@ -4,34 +4,7 @@ open FSharp.CosmosDb
 open FSharp.Control
 open Microsoft.Extensions.Configuration
 open System.IO
-
-type Parent =
-    { FamilyName: string
-      FirstName: string }
-
-type Pet = { GivenName: string }
-
-type Child =
-    { FamilyName: string
-      FirstName: string
-      Gender: string
-      Grade: int
-      Pets: Pet array }
-
-type Address =
-    { State: string
-      Country: string
-      City: string }
-
-type Family =
-    { [<Id>]
-      Id: string
-      [<PartitionKey>]
-      LastName: string
-      IsRegistered: bool
-      Parents: Parent array
-      Children: Child array
-      Address: Address }
+open Types
 
 let getFamiliesConnection host key =
     host
@@ -110,6 +83,8 @@ let main argv =
 
         // let conn =
         //     getFamiliesConnectionFromConnString connectionString
+
+        printfn "Getting ready to do some Cosmos operations"
 
         let families =
             [| { Id = "Powell.1"